// src/components/gdpr/types.ts
export interface GDPRConfig {
  findOut?: {
    text: string;
    link: string;
  };
  banner?: {
    heading: string;
    text: string;
    btnOk: string;
    btnEdit: string;
  };
  modal?: {
    heading: string;
    text: string;
    btnOk: string;
    btnSave: string;
    btnDeny: string;
    rHeading: string;
    rText: string;
    fHeading: string;
    fText: string;
    aHeading: string;
    aText: string;
  };
}